refactor: simplify DukanStatusContent and PendingDukanScreen
