--- dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/dukanStatus/PendingDukan.kt
@@ -1,6 +1,0 @@
-package net.thechance.mena.dukan.presentation.screen.pendingDukan
-
-enum class PendingDukan() {
-    PREFIX_TITLE,
-    SUFFIX_TITLE
-}
