+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/pendingDukan/PendingDukanScreen.kt
@@ -4,18 +4,25 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.withStyle
 import mena.dukan_presentation.generated.resources.Res
+import mena.dukan_presentation.generated.resources.dukan_blur
+import mena.dukan_presentation.generated.resources.dukan_pending
+import mena.dukan_presentation.generated.resources.dukan_request_pending
+import mena.dukan_presentation.generated.resources.dukan_waiting_approval
 import mena.dukan_presentation.generated.resources.ic_arrow_left
 import mena.dukan_presentation.generated.resources.my_dukan
 import net.thechance.mena.designsystem.presentation.component.appBar.AppBar
 import net.thechance.mena.designsystem.presentation.component.icon.Icon
 import net.thechance.mena.designsystem.presentation.theme.theme.MenaTheme
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
-import net.thechance.mena.dukan.presentation.viewModel.mainScreen.MainScreenUiState
+import net.thechance.mena.dukan.presentation.component.AnnotatedText
+import net.thechance.mena.dukan.presentation.component.ImageWithTextContainer
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 import org.jetbrains.compose.ui.tooling.preview.Preview
@@ -23,7 +30,6 @@ import org.jetbrains.compose.ui.tooling.preview.Preview
 @Composable
 fun PendingDukanScreen(
     dukanName: String,
-    dukanStatus: MainScreenUiState.DukanStatusUi,
     onBackClick: () -> Unit,
 ) {
     Column(
@@ -45,18 +51,43 @@ fun PendingDukanScreen(
 
         Spacer(modifier = Modifier.weight(1f))
 
-        DukanStatusContent(
-            dukanState = MainScreenUiState.DukanState(
-                name = dukanName,
-                status = dukanStatus
-            ),
-            modifier = Modifier.fillMaxWidth()
+        val titleText = buildPendingDukanTitle(
+            brandName = dukanName,
+            titleTemplate = stringResource(Res.string.dukan_request_pending)
+        )
+
+        ImageWithTextContainer(
+            blurImageRes = Res.drawable.dukan_blur,
+            foregroundImageRes = Res.drawable.dukan_pending,
+            header = {
+                AnnotatedText(
+                    text = titleText,
+                    style = TextStyle(textAlign = TextAlign.Center)
+                )
+            },
+            bodyText = stringResource(Res.string.dukan_waiting_approval),
         )
 
         Spacer(modifier = Modifier.weight(1f))
     }
 }
 
+@Composable
+private fun buildPendingDukanTitle(
+    brandName: String,
+    titleTemplate: String,
+) = androidx.compose.ui.text.buildAnnotatedString {
+    val parts = titleTemplate.split("%s")
+    withStyle(Theme.typography.title.small.toSpanStyle()) {
+        append(parts.getOrElse(0) { "" })
+    }
+    withStyle(Theme.typography.title.medium.toSpanStyle()) {
+        append(brandName)
+    }
+    withStyle(Theme.typography.title.small.toSpanStyle()) {
+        append(parts.getOrElse(1) { "" })
+    }
+}
 
 @Preview
 @Composable
@@ -65,7 +96,6 @@ private fun Preview() {
         PendingDukanScreen(
             dukanName = "Calvin Klein",
             onBackClick = {},
-            dukanStatus = MainScreenUiState.DukanStatusUi.Approved
         )
     }
 }
