+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/viewModel/mainScreen/MainViewModel.kt
@@ -55,7 +55,7 @@ class MainViewModel(
         when (state.value.dukanState.status) {
             DukanStatusUi.None -> emitEffect(MainEffect.NavigateToAddDukanScreen)
             DukanStatusUi.Pending -> emitEffect(MainEffect.NavigateToPendingDukanScreen)
-            DukanStatusUi.Approved -> emitEffect(MainEffect.NavigateToPendingDukanScreen)
+            DukanStatusUi.Approved -> TODO()
         }
     }
 }
